pyproject.toml
[project]
name = "django-stripe-payment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "dotenv>=0.9.9",
    "lint>=1.2.1",
    "python-stripe>=0.0.6",
    "ruff>=0.12.7",
]
